service/requirements.txt
flask
openai
python-dotenv
regex